+++ laktory/models/sql/column.py
@@ -38,6 +38,7 @@ class SparkFuncArg(BaseModel):
 
     def to_spark(self):
         import pyspark.sql.functions as F
+
         v = self.value
         if self.to_column:
             v = F.expr(v)
@@ -139,9 +140,11 @@ class Column(BaseModel):
     # ----------------------------------------------------------------------- #
 
     def to_spark(
-        self, df, udfs: list[Callable[[...], SparkColumn]] = None
+        self,
+        df,
+        udfs: list[Callable[[...], SparkColumn]] = None,
+        raise_exception: bool = True,
     ) -> SparkColumn:
-
         import pyspark.sql.functions as F
         from laktory.spark import functions as LF
         from laktory.spark.dataframe import has_column
@@ -211,9 +214,13 @@ class Column(BaseModel):
             args += [_arg.to_spark()]
 
         if expected_cols > 0 and found_cols == 0:
-            raise ValueError(
-                f"None of the inputs columns ({_args}) for {self.name} have been found"
-            )
+            if raise_exception:
+                raise ValueError(
+                    f"None of the inputs columns ({_args}) for {self.name} have been found"
+                )
+            else:
+                logger.info("Input columns not available. Skipping")
+                return None
 
         # Build kwargs
         kwargs = {}