+++ tests/test_dataevent.py
@@ -4,7 +4,7 @@ import pytest
 
 from laktory._testing import StockPriceDataEventHeader
 from laktory._testing import EventsManager
-
+from laktory import models
 
 header = StockPriceDataEventHeader()
 events = EventsManager().build_events()
@@ -59,6 +59,7 @@ def test_model_dump():
             "_created_at": "2023-09-01T00:00:00Z",
         },
         "tstamp_col": "created_at",
+        "tstamp_in_path": True,
     }
 
     # With exclusions
@@ -81,6 +82,16 @@ def test_model_dump():
             "_created_at": "2023-09-01T00:00:00Z",
         },
     }
+
+
+def test_event_without_tstamp():
+    d = event.model_dump()
+    d["tstamp_in_path"] = False
+    e = models.DataEvent(**d)
+    assert (
+        e.get_landing_filepath()
+        == "/Volumes/dev/sources/landing/events/yahoo-finance/stock_price/stock_price.json"
+    )
 
 
 def test_to_azure_storage_container():
@@ -117,6 +128,7 @@ if __name__ == "__main__":
     test_dataeventheader()
     test_dataevent()
     test_model_dump()
+    test_event_without_tstamp()
     test_to_azure_storage_container()
     test_to_aws_s3_bucket()
     test_to_databricks_mount()